+++ glfw/backend_utils.c
@@ -222,9 +222,23 @@ dispatchTimers(EventLoopData *eld) {
 }
 
 static void
-drain_wakeup_fd(int fd, int events UNUSED, void* data UNUSED) {
+drain_wakeup_fd(int fd, EventLoopData* eld) {
     static char drain_buf[64];
-    while(read(fd, drain_buf, sizeof(drain_buf)) < 0 && errno == EINTR);
+    eld->wakeup_data_read = false;
+    while(true) {
+        ssize_t ret = read(fd, drain_buf, sizeof(drain_buf));
+        if (ret < 0) {
+            if (errno == EINTR) continue;
+            break;
+        }
+        if (ret > 0) { eld->wakeup_data_read = true; continue; }
+        break;
+    }
+}
+
+static void
+mark_wakep_fd_ready(int fd UNUSED, int events UNUSED, void *data) {
+    ((EventLoopData*)(data))->wakeup_fd_ready = true;
 }
 
 bool
@@ -238,17 +252,27 @@ initPollData(EventLoopData *eld, int display_fd) {
     if (pipe2(eld->wakeupFds, O_CLOEXEC | O_NONBLOCK) != 0) return false;
     const int wakeup_fd = eld->wakeupFds[0];
 #endif
-    if (!addWatch(eld, "wakeup", wakeup_fd, POLLIN, 1, drain_wakeup_fd, NULL)) return false;
+    if (!addWatch(eld, "wakeup", wakeup_fd, POLLIN, 1, mark_wakep_fd_ready, eld)) return false;
     return true;
+}
+
+void
+check_for_wakeup_events(EventLoopData *eld) {
+#ifdef HAS_EVENT_FD
+    int fd = eld->wakeupFd;
+#else
+    int fd = eld->wakeupFds[0];
+#endif
+    drain_wakeup_fd(fd, eld);
 }
 
 void
 wakeupEventLoop(EventLoopData *eld) {
 #ifdef HAS_EVENT_FD
-    static const int64_t value = 1;
-    while (write(eld->wakeupFd, &value, sizeof value) < 0 && errno == EINTR);
+    static const uint64_t value = 1;
+    while (write(eld->wakeupFd, &value, sizeof value) < 0 && (errno == EINTR || errno == EAGAIN));
 #else
-    while (write(eld->wakeupFds[1], "w", 1) < 0 && errno == EINTR);
+    while (write(eld->wakeupFds[1], "w", 1) < 0 && (errno == EINTR || errno == EAGAIN));
 #endif
 }
 
@@ -281,6 +305,7 @@ pollForEvents(EventLoopData *eld, double timeout) {
     EVDBG("pollForEvents final timeout: %.3f", timeout);
     int result;
     double end_time = monotonic() + timeout;
+    eld->wakeup_fd_ready = false;
 
     while(1) {
         if (timeout >= 0) {